Time: 700 ms (27.81%), Space: 86.1 MB (26.65%) - LeetHub

+++ 1465-maximum-product-of-splitted-binary-tree/1465-maximum-product-of-splitted-binary-tree.cpp
@@ -24,8 +24,8 @@ public:
     }
 
     int maxProduct(TreeNode* root) {
-        int ans=0;
-        int sum=FindSum(root);
+        long long ans=0;
+        long long sum=FindSum(root);
         for(auto i:sums){
             ans=max(ans,i*(sum-i));
         }
